(feat) No opts on locked lists

+++ src/db/deleteItem.ts
@@ -1,25 +1,41 @@
 import { db } from '../firebase';
-import { QueryRef, DbFunc, ListItem } from '../types/db';
+import { QueryRef, DbFunc, ListItem, DocRef, List } from '../types/db';
 import { msgFormat } from '../utils/discord';
 
 export const deleteItem: DbFunc<Opts, string> = async opts => {
-  const query = db
-    .collection(`owners/${opts.ownerID}/lists/${opts.list}/items`)
-    .orderBy('createdAt')
+  const listRef = db.doc(`owners/${opts.ownerID}/lists/${opts.list}`) as DocRef<List>;
+  const itemQuery = listRef
+    .collection(`items`)
+    .orderBy('rank')
     .offset(opts.itemNum - 1)
     .limit(1) as QueryRef<ListItem>;
 
-  const itemDocs = (await query.get()).docs;
+  return db.runTransaction(async t => {
+    const list = (await t.get(listRef)).data();
 
-  if (!itemDocs.length) {
-    return { success: false, message: 'Invalid item' };
-  }
+    if (!list) {
+      return { success: false, message: `You don't have a list named ${msgFormat.code(opts.list)}.` };
+    }
 
-  const item = itemDocs[0].data();
+    if (list.isLocked) {
+      return {
+        success: false,
+        message: `This list is currently locked for normalization. Please try again in a moment`,
+      };
+    }
 
-  await itemDocs[0].ref.delete();
+    const itemDocs = (await t.get(itemQuery)).docs;
 
-  return { success: true, data: `${msgFormat.code(item.name)} has been deleted from ${msgFormat.code(opts.list)}.` };
+    if (!itemDocs.length) {
+      return { success: false, message: 'Invalid item' };
+    }
+
+    const item = itemDocs[0].data();
+
+    await itemDocs[0].ref.delete();
+
+    return { success: true, data: `${msgFormat.code(item.name)} has been deleted from ${msgFormat.code(opts.list)}.` };
+  });
 };
 
 interface Opts {
